✨ Feat :  사진 미리보기 화면을 인앱 브라우저 화면으로 수정하여 적용
